feat: featured posts and split carousels into components
