--- src/lib/components/carousel/index.ts
@@ -1,1 +1,0 @@
-export { default as Carousel } from './carousel.svelte';
