+++ src/lib/components/carousels/carousel-brands/index.ts
@@ -1,0 +1,1 @@
+export { default as CarouselBrands } from './carousel-brands.svelte';
