+++ src/lib/components/carousels/carousel-featured-posts/index.ts
@@ -1,0 +1,1 @@
+export { default as CarouselFeaturedPosts } from './carousel-featured-posts.svelte';
