+++ src/lib/components/carousels/carousel-projects/index.ts
@@ -1,0 +1,1 @@
+export { default as CarouselProjects } from './carousel-projects.svelte';
