+++ src/lib/components/carousels/index.ts
@@ -1,0 +1,3 @@
+export * from './carousel-brands';
+export * from './carousel-featured-posts';
+export * from './carousel-projects';
