+++ src/lib/ts/interfaces/project.ts
@@ -1,0 +1,8 @@
+export interface IProject {
+	title: string;
+	description: string;
+	previewImage: string;
+	repo: string;
+	deploy: string;
+	tags: string[];
+}
